# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Africa Travel Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  fullName: string;
  nationality: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  fullName: string;
  nationality: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Country {
  id: number;
  name: string;
  /** @nullable */
  code?: string | null;
  type: string;
  createdAt: string;
}

export interface CountryInput {
  name: string;
  code?: string;
}

export interface Tour {
  id: number;
  title: string;
  description: string;
  durationDays: number;
  destinationCountryId: number;
  basePrice: number;
  highlights: string[];
  included: string[];
  notIncluded: string[];
  createdAt: string;
  updatedAt: string;
}

export interface TourSummary {
  id: number;
  title: string;
  description: string;
  durationDays: number;
  destinationCountryId: number;
  basePrice: number;
  highlights?: string[];
  /** @nullable */
  coverImage?: string | null;
  destinationCountry?: Country | null;
  createdAt: string;
  updatedAt: string;
}

export interface TourImage {
  id: number;
  tourId: number;
  imageUrl: string;
  /** @nullable */
  caption?: string | null;
  sortOrder: number;
}

export interface TourPricing {
  id: number;
  tourId: number;
  originCountryId: number;
  price: number;
}

export interface TourDetail {
  id: number;
  title: string;
  description: string;
  durationDays: number;
  destinationCountryId: number;
  basePrice: number;
  highlights: string[];
  included: string[];
  notIncluded: string[];
  images: TourImage[];
  pricing: TourPricing[];
  destinationCountry?: Country | null;
  createdAt: string;
  updatedAt: string;
}

export interface TourInput {
  title: string;
  description: string;
  durationDays: number;
  destinationCountryId: number;
  basePrice: number;
  highlights?: string[];
  included?: string[];
  notIncluded?: string[];
}

export interface TourUpdate {
  title?: string;
  description?: string;
  durationDays?: number;
  destinationCountryId?: number;
  basePrice?: number;
  highlights?: string[];
  included?: string[];
  notIncluded?: string[];
}

export interface TourImageInput {
  imageUrl: string;
  caption?: string;
  sortOrder?: number;
}

export interface TourPricingInput {
  originCountryId: number;
  price: number;
}

export interface Booking {
  id: number;
  userId: number;
  tourId: number;
  originCountryId: number;
  numberOfPeople: number;
  totalPrice: number;
  status: string;
  bookingDate: string;
  /** @nullable */
  paymentDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface BookingDetail {
  id: number;
  userId: number;
  tourId: number;
  originCountryId: number;
  numberOfPeople: number;
  totalPrice: number;
  status: string;
  bookingDate: string;
  /** @nullable */
  paymentDate?: string | null;
  /** @nullable */
  notes?: string | null;
  tour?: TourSummary | null;
  user?: User | null;
  originCountry?: Country | null;
}

export interface BookingInput {
  tourId: number;
  originCountryId: number;
  numberOfPeople: number;
  notes?: string;
}

export interface BookingStatusUpdate {
  status: string;
}

export interface VisaService {
  id: number;
  name: string;
  description: string;
  fee: number;
  destinationCountryId: number;
  requirements: string[];
  isActive: boolean;
  destinationCountry?: Country | null;
  createdAt: string;
}

export interface VisaServiceInput {
  name: string;
  description: string;
  fee: number;
  destinationCountryId: number;
  requirements?: string[];
  isActive?: boolean;
}

export interface VisaServiceUpdate {
  name?: string;
  description?: string;
  fee?: number;
  destinationCountryId?: number;
  requirements?: string[];
  isActive?: boolean;
}

export interface VisaCase {
  id: number;
  userId: number;
  visaServiceId: number;
  status: string;
  feePaid: boolean;
  /** @nullable */
  submittedAt?: string | null;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
}

export interface CaseDocument {
  id: number;
  visaCaseId: number;
  documentType: string;
  fileUrl: string;
  uploadedAt: string;
}

export interface VisaCaseDetail {
  id: number;
  userId: number;
  visaServiceId: number;
  status: string;
  feePaid: boolean;
  /** @nullable */
  submittedAt?: string | null;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
  service?: VisaService | null;
  documents?: CaseDocument[];
  user?: User | null;
}

export interface VisaCaseInput {
  visaServiceId: number;
}

export interface VisaCaseStatusUpdate {
  status: string;
  adminNotes?: string;
}

export interface CaseDocumentInput {
  documentType: string;
  fileUrl: string;
}

export interface GalleryImage {
  id: number;
  imageUrl: string;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  countryId?: number | null;
  country?: Country | null;
  sortOrder: number;
  createdAt: string;
}

export interface GalleryImageInput {
  imageUrl: string;
  caption?: string;
  countryId?: number;
  sortOrder?: number;
}

export interface AdminStats {
  totalUsers: number;
  totalBookings: number;
  totalRevenue: number;
  pendingVisaCases: number;
  totalTours: number;
  confirmedBookings: number;
  cancelledBookings: number;
}

export type ListToursParams = {
destinationCountryId?: number;
maxPrice?: number;
minDays?: number;
maxDays?: number;
};

export type GetAllBookingsParams = {
status?: string;
};

export type GetAllVisaCasesParams = {
status?: string;
};