# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Africa Travel Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Countries (public) ────────────────────────────────────────────────────
...
  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking (user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /bookings/{id}/mark-paid:
    post:
      operationId: markBookingPaid
      tags: [bookings]
      summary: Mark booking as paid (user)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking marked paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/admin/all:
    get:
      operationId: getAllBookings
      tags: [bookings]
      summary: Get all bookings (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: All bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingDetail"
  /bookings/admin/{id}:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # ── Visa ──────────────────────────────────────────────────────────────────
...
  /admin/countries/origin:
    get:
      operationId: adminListOriginCountries
      tags: [admin]
      summary: List origin countries (admin)
      responses:
        "200":
          description: Origin countries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Country"
    post:
      operationId: addOriginCountry
      tags: [admin]
      summary: Add origin country (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CountryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Country"
  /admin/countries/destination:
    get:
      operationId: adminListDestinationCountries
      tags: [admin]
      summary: List destination countries (admin)
      responses:
        "200":
          description: Destination countries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Country"
    post:
      operationId: addDestinationCountry
      tags: [admin]
      summary: Add destination country (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CountryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Country"
  /admin/countries/{id}:
    delete:
      operationId: deleteCountry
      tags: [admin]
      summary: Delete a country (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Gallery ───────────────────────────────────────────────────────────────
...
  /admin/gallery:
    get:
      operationId: getAdminGallery
      tags: [admin]
      summary: All gallery images (admin)
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"
    post:
      operationId: addGalleryImage
      tags: [admin]
      summary: Add gallery image (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryImageInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
  /admin/gallery/{id}:
    delete:
      operationId: deleteGalleryImage
      tags: [admin]
      summary: Delete gallery image (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Auth ────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [email, password, fullName, nationality]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
        fullName:
          type: string
        nationality:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, email, fullName, nationality, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        nationality:
          type: string
        role:
          type: string
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ── Countries ───────────────────────────────────────────────────────────
    Country:
      type: object
      required: [id, name, type, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: ["string", "null"]
        type:
          type: string
        createdAt:
          type: string
    CountryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        code:
          type: string

    # ── Tours ───────────────────────────────────────────────────────────────
...
    TourSummary:
      type: object
      required: [id, title, description, durationDays, destinationCountryId, basePrice, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        durationDays:
          type: integer
        destinationCountryId:
          type: integer
        basePrice:
          type: number
        highlights:
          type: array
          items:
            type: string
        coverImage:
          type: ["string", "null"]
        destinationCountry:
          oneOf:
            - $ref: "#/components/schemas/Country"
            - type: "null"
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Booking:
      type: object
      required: [id, userId, tourId, originCountryId, numberOfPeople, totalPrice, status, bookingDate]
      properties:
        id:
          type: integer
        userId:
          type: integer
        tourId:
          type: integer
        originCountryId:
          type: integer
        numberOfPeople:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
        bookingDate:
          type: string
        paymentDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    BookingDetail:
      type: object
      required: [id, userId, tourId, originCountryId, numberOfPeople, totalPrice, status, bookingDate]
      properties:
        id:
          type: integer
        userId:
          type: integer
        tourId:
          type: integer
        originCountryId:
          type: integer
        numberOfPeople:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
        bookingDate:
          type: string
        paymentDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tour:
          oneOf:
            - $ref: "#/components/schemas/TourSummary"
            - type: "null"
        user:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        originCountry:
          oneOf:
            - $ref: "#/components/schemas/Country"
            - type: "null"
    BookingInput:
      type: object
      required: [tourId, originCountryId, numberOfPeople]
      properties:
        tourId:
          type: integer
        originCountryId:
          type: integer
        numberOfPeople:
          type: integer
        notes:
          type: string
    BookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string

    # ── Visa ────────────────────────────────────────────────────────────────
...
    GalleryImage:
      type: object
      required: [id, imageUrl, sortOrder, createdAt]
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        countryId:
          type: ["integer", "null"]
        country:
          oneOf:
            - $ref: "#/components/schemas/Country"
            - type: "null"
        sortOrder:
          type: integer
        createdAt:
          type: string
    GalleryImageInput:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
        caption:
          type: string
        countryId:
          type: integer
        sortOrder:
          type: integer

    # ── Admin Stats ─────────────────────────────────────────────────────────
...