# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Africa Travel Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  userId: number;
  tourId: number;
  originCountryId: number;
  numberOfPeople: number;
  totalPrice: number;
  status: string;
  bookingDate: string;
  /** @nullable */
  paymentDate?: string | null;
  /** @nullable */
  notes?: string | null;
}