# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Africa Travel Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  fullName: string;
  nationality: string;
  role: string;
  createdAt: string;
}